+++ src/SickoSampler.cpp
@@ -1011,6 +1011,64 @@ struct SickoSampler : Module {
 		INFO ("[ SickoCV ] RESAMPLED");
 		*/
 	}	
+
+
+//	-----------------------------------------------------------------------------------------------
+
+	float* LoadWavFileF32(const std::string& path, uint32_t* channels, uint32_t* sampleRate, uint64_t* totalSampleCount) {
+	    drwav wav;
+	    if (!drwav_init_file(&wav, path.c_str(), nullptr)) {
+	        return nullptr;
+	    }
+
+	    if (channels) *channels = wav.channels;
+	    if (sampleRate) *sampleRate = wav.sampleRate;
+
+	    uint64_t frameCount = wav.totalPCMFrameCount;
+	    uint64_t sampleCount = frameCount * wav.channels;
+
+	    float* pSampleData = (float*)malloc((size_t)sampleCount * sizeof(float));
+	    if (!pSampleData) {
+	        drwav_uninit(&wav);
+	        return nullptr;
+	    }
+
+	    uint64_t framesRead = drwav_read_pcm_frames_f32(&wav, frameCount, pSampleData);
+	    drwav_uninit(&wav);
+
+	    if (totalSampleCount) *totalSampleCount = framesRead * wav.channels;
+
+	    return pSampleData;
+	}
+
+// -------------------------------------------------------------------------------------------------------------------------------
+
+	bool SaveWavFileF32(const std::string& path, const std::vector<float>& data, uint32_t sampleRate, uint32_t channels) {
+	    drwav_data_format format;
+	    format.container = drwav_container_riff;      // Standard WAV
+	    format.format = DR_WAVE_FORMAT_IEEE_FLOAT;    // Float 32-bit
+	    format.channels = channels;
+	    format.sampleRate = sampleRate;
+	    format.bitsPerSample = 32;
+
+	    drwav wav;
+	    if (!drwav_init_file_write(&wav, path.c_str(), &format, nullptr)) {
+	        return false;
+	    }
+
+	    drwav_uint64 framesToWrite = data.size() / channels;
+
+	    // Scrivi i frame
+	    drwav_uint64 framesWritten = drwav_write_pcm_frames(&wav, framesToWrite, data.data());
+
+	    drwav_uninit(&wav);
+
+	    return framesWritten == framesToWrite;
+	}
+
+// -------------------------------------------------------------------------------------------------------------------------------
+
+
 /*
 
 																							░██████╗░█████╗░██╗░░░██╗███████╗
@@ -1122,9 +1180,16 @@ struct SickoSampler : Module {
 		if (path.substr(path.size() - 4) != ".wav" && path.substr(path.size() - 4) != ".WAV")
 			path += ".wav";
 
+		/*
 		drwav *pWav = drwav_open_file_write(path.c_str(), &format);
 		drwav_write(pWav, samples, data.data());
 		drwav_close(pWav);
+		*/
+		bool ok = SaveWavFileF32(path.c_str(), data, format.sampleRate, format.channels);
+		if (!ok) {
+		    // std::cerr << "Errore durante il salvataggio WAV" << std::endl;
+		    INFO("ERROR WRITING");
+		}
 
 		char* pathDup = strdup(path.c_str());
 
@@ -1242,8 +1307,10 @@ struct SickoSampler : Module {
 		unsigned int c;
 		unsigned int sr;
 		drwav_uint64 tsc;
-		float* pSampleData;
-		pSampleData = drwav_open_and_read_file_f32(path.c_str(), &c, &sr, &tsc);
+		//float* pSampleData;
+		//pSampleData = drwav_open_and_read_file_f32(path.c_str(), &c, &sr, &tsc);
+		float* pSampleData = LoadWavFileF32(path.c_str(), &c, &sr, &tsc);	// new dr_wav lib
+
 
 		if (pSampleData != NULL && tsc > minSamplesToLoad * c) {
 			fileFound = true;
@@ -1284,7 +1351,7 @@ struct SickoSampler : Module {
 				}
 				totalSampleC = playBuffer[LEFT][0].size();
 				totalSamples = totalSampleC-1;
-				drwav_free(pSampleData);
+//				drwav_free(pSampleData);
 
 				for (unsigned int i = 1; i < totalSamples; i = i+2) {
 					playBuffer[LEFT][0][i] = playBuffer[LEFT][0][i-1] * .5f + playBuffer[LEFT][0][i+1] * .5f;
@@ -1307,7 +1374,7 @@ struct SickoSampler : Module {
 				}
 				totalSampleC = playBuffer[LEFT][0].size();
 				totalSamples = totalSampleC-1;
-				drwav_free(pSampleData);
+//				drwav_free(pSampleData);
 
 				resampled = false;
 
@@ -1323,7 +1390,7 @@ struct SickoSampler : Module {
 					}
 				}
 
-				drwav_free(pSampleData);
+//				drwav_free(pSampleData);
 
 				drwav_uint64 tempSampleC = tempBuffer[LEFT].size();
 				drwav_uint64 tempSamples = tempSampleC-1;